refactor: DispatcherServletInitializer를 mvc 모듈로 이동
