+++ mvc/src/main/java/com/interface21/webmvc/servlet/DispatcherServlet.java
@@ -19,15 +19,14 @@ public class DispatcherServlet extends HttpServlet {
     private static final Logger log = LoggerFactory.getLogger(DispatcherServlet.class);
     private static final HandlerMappingRegistry handlerMappingRegistry = new HandlerMappingRegistry();
     private static final HandlerAdapterRegistry handlerAdapterRegistry = new HandlerAdapterRegistry();
-    private final String basePackage;
 
-    public DispatcherServlet(String basePackage) {
-        this.basePackage = basePackage;
+    public DispatcherServlet() {
+
     }
 
     @Override
     public void init() {
-        handlerMappingRegistry.addHandlerMapping(new AnnotationHandlerMapping(basePackage));
+        handlerMappingRegistry.addHandlerMapping(new AnnotationHandlerMapping());
         handlerAdapterRegistry.addHandlerAdapter(new AnnotationHandlerAdapter());
     }
