+++ mvc/src/main/java/com/interface21/webmvc/servlet/DispatcherServletInitializer.java
@@ -1,7 +1,6 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet;
 
 import com.interface21.web.WebApplicationInitializer;
-import com.interface21.webmvc.servlet.DispatcherServlet;
 import jakarta.servlet.ServletContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -15,11 +14,10 @@ public class DispatcherServletInitializer implements WebApplicationInitializer {
     private static final Logger log = LoggerFactory.getLogger(DispatcherServletInitializer.class);
 
     private static final String DEFAULT_SERVLET_NAME = "dispatcher";
-    private static final String BASE_PACKAGE = "com.techcourse";
 
     @Override
     public void onStartup(final ServletContext servletContext) {
-        final var dispatcherServlet = new DispatcherServlet(BASE_PACKAGE);
+        final var dispatcherServlet = new DispatcherServlet();
 
         final var registration = servletContext.addServlet(DEFAULT_SERVLET_NAME, dispatcherServlet);
         if (registration == null) {
